Moves renderErrorMessage into shared folder
