+++ src/components/form/AddGeneticsModalContent.tsx
@@ -18,7 +18,7 @@ import { RootState } from '../../redux/store'
 import { addGeneticsSampleSchema } from '../../utils/helpers/yupValidations'
 import CustomModalHeader from '../Shared/CustomModalHeader'
 import CustomSelect from '../Shared/CustomSelect'
-import renderErrorMessage from './RenderErrorMessage'
+import renderErrorMessage from '../Shared/RenderErrorMessage'
 
 const initialFormValues = {
   sampleIdNumber: '',
