+++ src/components/form/MarkFishModalContent.tsx
@@ -17,7 +17,7 @@ import { RootState } from '../../redux/store'
 import { addMarksOrTagsSchema } from '../../utils/helpers/yupValidations'
 import CustomModalHeader from '../Shared/CustomModalHeader'
 import CustomSelect from '../Shared/CustomSelect'
-import renderErrorMessage from './RenderErrorMessage'
+import renderErrorMessage from '../Shared/RenderErrorMessage'
 
 const initialFormValues = {
   type: '',
